added: logos, splash, homepage redesign
