+++ App.js
@@ -9,47 +9,15 @@
 import 'react-native-gesture-handler';
 import React, {useEffect, useState} from 'react';
 import {StatusBar, useColorScheme, StyleSheet, View, Text} from 'react-native';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-
-import {NavigationContainer} from '@react-navigation/native';
-import {createStackNavigator} from '@react-navigation/stack';
-
-import LoginDrawerNavigator from './navigation/LoginDrawerNavigation';
-
 import HomePage from './screens/HomePage';
-import Login from './screens/Login';
-import Signup from './screens/Signup';
-import About from './screens/About';
-import ContactUs from './screens/ContactUs';
-
-const Stack = createStackNavigator();
 
 const App = () => {
   const isDarkMode = useColorScheme() === 'dark';
 
   const backgroundStyle = {
     backgroundColor: isDarkMode
-      ? styles.backgroundColorTheme.dark
-      : styles.backgroundColorTheme.light,
-  };
-
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-
-  useEffect(() => {
-    getData();
-  }, []);
-
-  const getData = () => {
-    try {
-      AsyncStorage.getItem('userName').then(value => {
-        if (value != null) {
-          setIsLoggedIn(true);
-          console.log({value});
-        }
-      });
-    } catch (error) {
-      console.log({error});
-    }
+      ? styles.backgroundThemeColor.dark
+      : styles.backgroundThemeColor.light,
   };
 
   return (
@@ -58,19 +26,7 @@ const App = () => {
         barStyle={isDarkMode ? 'light-content' : 'dark-content'}
         backgroundColor={backgroundStyle.backgroundColor}
       />
-      <NavigationContainer>
-        <Stack.Navigator screenOptions={{headerShown: false}}>
-          <Stack.Screen name="HomePage" component={HomePage} />
-          <Stack.Screen name="Login" component={Login} />
-          <Stack.Screen name="Signup" component={Signup} />
-          <Stack.Screen name="About" component={About} />
-          <Stack.Screen name="ContactUs" component={ContactUs} />
-          <Stack.Screen
-            name="LoginDrawerNavigator"
-            component={LoginDrawerNavigator}
-          />
-        </Stack.Navigator>
-      </NavigationContainer>
+      <HomePage />
     </View>
   );
 };
@@ -79,15 +35,14 @@ const styles = StyleSheet.create({
   body: {
     flex: 1,
   },
-  text: {
-    color: '#0F0F0F',
+  textSize: {
     fontSize: 40,
   },
-  backgroundColorTheme: {
-    dark: '#00155F',
+  backgroundThemeColor: {
+    dark: '#1c1e21',
     light: '#f0f0ed',
   },
-  textColor: {
+  themeTextColor: {
     dark: '#000000',
     light: '#FFFFFF',
   },
